Move UnknownBaseContractIdError to where it's used
- Remove from lib/types/errors, add to lib/allocate/storage
- Also remove an unused import
- Update lib/index to reflect new location
